Add intent filter to start the app on a GPX file from the Android file manager
In this way, the normal file manager can be used to start this app to
load the file.

It is needed, however, to split loading a file and starting playback
into two separate operations. This requires quite some changes.
The split is needed to give the user control over starting the GPX
playback separate from the loading step.

One thing to note is that the Binder connection is not there yet in
onCreate, so the data Uri from the intent needs to be saved until that
connection is there.